+++ src/main/java/com/airbnb/crud/airbnbDB/customer/repository/CustomerRWRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CustomerRWRepository extends JpaRepository<Customer, Integer> {
+public interface CustomerRWRepository extends JpaRepository<Customer, Long> {
 }
