+++ src/main/java/com/airbnb/crud/airbnbDB/host/repository/HostRWRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HostRWRepository extends JpaRepository<Host, Integer> {
+public interface HostRWRepository extends JpaRepository<Host, Long> {
 }
